Krishibodh-Robot-Core/enhanced_app/requirements.txt
flask
pyserial
requests
google-generativeai
opencv-python
rembg
pillow
onnxruntime
numpy
